requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
xgboost==2.0.2
optuna==3.4.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
matplotlib==3.8.0
seaborn==0.13.0
plotly==5.17.0
kaleido==0.2.1
